Move QuickActions out of lib/
