+++ .eslintignore
@@ -51,5 +51,5 @@ ReactNativeClient/pluginAssets/
 # Ignore files generated from TypeScript files
 ElectronClient/app/gui/ShareNoteDialog.js
 ReactNativeClient/lib/JoplinServerApi.js
-ReactNativeClient/lib/QuickActions.js
+ReactNativeClient/QuickActions.js
 ReactNativeClient/pluginAssetsLoader.js
