+++ .gitignore
@@ -49,5 +49,5 @@ Tools/commit_hook.txt
 # Ignore files generated from TypeScript files
 ElectronClient/app/gui/ShareNoteDialog.js
 ReactNativeClient/lib/JoplinServerApi.js
-ReactNativeClient/lib/QuickActions.js
+ReactNativeClient/QuickActions.js
 ReactNativeClient/pluginAssetsLoader.js
